Build bundle.js with Webpack
We initialize a new npm package:

`$ npm init`

We remove the `main` field from `package.json` because this application is
not a module for others to depend on. As long as there are no meaningful
tests, the `test` script is left to personal preference. I like `false` as
a reminder to write tests.

We add and configure Webpack to compile our JavaScript sources into a bundle.

`$ yarn add webpack --dev`

https://webpack.js.org/guides/get-started/#using-webpack-with-a-config

We add a `build` script to `package.json` to invoke Webpack. Now we can build
our bundle with `npm run build`.

We tell Git to ignore the Webpack build directory (`/dist`) and the Node.js
modules directory (`/node_modules`) in `.gitignore`.