Serve application with webpack-dev-server
We add and configure webpack-dev-server:

`$ yarn add webpack-dev-server --dev`

https://webpack.js.org/guides/development/#webpack-dev-server
https://webpack.js.org/configuration/dev-server/

The server can serve both the bundles generated by Webpack and the static
files sitting in our working directory. The server keeps the bundles in
memory and never writes them to disk. Bundles are served relative to the path
configured by `devSever.publicPath`, which is `"/"` by default. That means if
a sever is running at `http://localhost:8080` and `output.filename` is
`bundle.js`, then the bundle is available at
`http://localhost:8080/bundle.js`. In our `index.html`, we reference it as
`/bundle.js`.

The server searches `devSever.contentBase` for static files, and they are
always served relative to the root path, `"/"`.

We add a `start` script to `package.json` that will start a server and open
a browser at its root path. We can run the script with `npm start`. Once
open, we can change `src/main.js` and the server will automatically rebuild
the bundle and refresh the page. If we change `src/index.html`, it will not
refresh unless we configure `devServer.watchContentBase` to `true`.
Generally, we want all of our application in the bundle served by a barebones
HTML skeleton page, so we will leave that option at its default, `false`.