Enable source maps
Configure TypeScript to attach source maps and Webpack to pass them through:

`$ yarn add --dev source-map-loader`

https://webpack.js.org/guides/webpack-and-typescript/#enabling-source-maps

When connecting to the development server, sources will appear to be served
from `webpack:///src/`. We can browse them and set breakpoints that work.
(Sadly, however, messages in the console will not be linked back to their
origin in the source.)