Use tslint
In the general case, a linter is a static analysis tool. Linters are commonly
used to check formatting and enforce best practices. Folding linters into the
general test suite is an effective way to maintain good code quality.

We add and configure tslint:

`$ yarn add --dev tslint`

https://github.com/palantir/tslint#configuration

We start with the "recommended" configuration. We can add per-project,
per-file, and per-line exceptions as we go.

https://github.com/palantir/tslint#rule-flags

There are more options for configurations we can explore later.

https://github.com/palantir/tslint#custom-rule-sets-from-the-community